Theme redesign, fixed mobile main menu
Changed the font across the entire app to a more simplistic and sleek font. The omnisearch has been revamped with a more simplistic and minimalist design with little color and form. Additionally, fixed the main menu ship display for mobile devices. It displays as a list instead of rows for mobile devices due to a rendering bug of some sort for rows.